ref(processor): Split off profile processor code into separate sub-module
