+++ relay-server/src/actors/processor.rs
@@ -24,15 +24,15 @@ use relay_event_normalization::{GeoIpLookup, RawUserAgentInfo};
 use relay_event_schema::processor::{self, process_value, ProcessingAction, ProcessingState};
 use relay_event_schema::protocol::{
     Breadcrumb, Contexts, Csp, Event, EventType, ExpectCt, ExpectStaple, Hpkp, IpAddr,
-    LenientString, Metrics, NetworkReportError, OtelContext, ProfileContext, RelayInfo,
-    SecurityReportType, Timestamp, TraceContext, Values,
+    LenientString, Metrics, NetworkReportError, OtelContext, RelayInfo, SecurityReportType,
+    Timestamp, TraceContext, Values,
 };
 use relay_filter::FilterStatKey;
 use relay_metrics::aggregator::partition_buckets;
 use relay_metrics::aggregator::AggregatorConfig;
 use relay_metrics::{Bucket, BucketsView, MergeBuckets, MetricMeta, MetricNamespace};
 use relay_pii::{scrub_graphql, PiiAttachmentsProcessor, PiiConfigError, PiiProcessor};
-use relay_profiling::{ProfileError, ProfileId};
+use relay_profiling::ProfileId;
 use relay_protocol::{Annotated, Array, Empty, FromValue, Object, Value};
 use relay_quotas::{DataCategory, Scoping};
 use relay_sampling::config::{RuleType, SamplingMode};
@@ -48,7 +48,7 @@ use tokio::sync::Semaphore;
 #[cfg(feature = "processing")]
 use {
     crate::actors::project_cache::UpdateRateLimits,
-    crate::utils::{EnvelopeLimiter, MetricsLimiter},
+    crate::utils::{EnvelopeLimiter, ItemAction, MetricsLimiter},
     relay_event_normalization::{span, StoreConfig, StoreProcessor},
     relay_event_schema::protocol::Span,
     relay_metrics::{Aggregator, RedisMetricMetaStore},
@@ -71,9 +71,10 @@ use crate::service::ServiceError;
 use crate::statsd::{PlatformTag, RelayCounters, RelayHistograms, RelayTimers};
 use crate::utils::{
     self, extract_transaction_count, ChunkedFormDataAggregator, ExtractionMode, FormDataIter,
-    ItemAction, ManagedEnvelope, SamplingResult,
+    ManagedEnvelope, SamplingResult,
 };
 
+mod profile;
 mod replay;
 mod report;
 mod session;
@@ -632,56 +633,6 @@ impl EnvelopeProcessorService {
         }
     }
 
-    /// Remove profiles from the envelope if they can not be parsed
-    fn filter_profiles(&self, state: &mut ProcessEnvelopeState) {
-        let transaction_count: usize = state
-            .managed_envelope
-            .envelope()
-            .items()
-            .filter(|item| item.ty() == &ItemType::Transaction)
-            .count();
-        let mut profile_id = None;
-        state.managed_envelope.retain_items(|item| match item.ty() {
-            // Drop profile without a transaction in the same envelope.
-            ItemType::Profile if transaction_count == 0 => ItemAction::DropSilently,
-            // First profile found in the envelope, we'll keep it if metadata are valid.
-            ItemType::Profile if profile_id.is_none() => {
-                match relay_profiling::parse_metadata(&item.payload(), state.project_id) {
-                    Ok(id) => {
-                        profile_id = Some(id);
-                        ItemAction::Keep
-                    }
-                    Err(err) => ItemAction::Drop(Outcome::Invalid(DiscardReason::Profiling(
-                        relay_profiling::discard_reason(err),
-                    ))),
-                }
-            }
-            // We found another profile, we'll drop it.
-            ItemType::Profile => ItemAction::Drop(Outcome::Invalid(DiscardReason::Profiling(
-                relay_profiling::discard_reason(ProfileError::TooManyProfiles),
-            ))),
-            _ => ItemAction::Keep,
-        });
-        state.profile_id = profile_id;
-    }
-
-    /// Transfers the profile ID from the profile item to the transaction item.
-    ///
-    /// If profile processing happens at a later stage, we remove the context again.
-    fn transfer_profile_id(&self, state: &mut ProcessEnvelopeState) {
-        if let Some(event) = state.event.value_mut() {
-            if event.ty.value() == Some(&EventType::Transaction) {
-                let contexts = event.contexts.get_or_insert_with(Contexts::new);
-                // If we found a profile, add its ID to the profile context on the transaction.
-                if let Some(profile_id) = state.profile_id {
-                    contexts.add(ProfileContext {
-                        profile_id: Annotated::new(profile_id),
-                    });
-                }
-            }
-        }
-    }
-
     /// Normalize monitor check-ins and remove invalid ones.
     #[cfg(feature = "processing")]
     fn process_check_ins(&self, state: &mut ProcessEnvelopeState) {
@@ -706,53 +657,6 @@ impl EnvelopeProcessorService {
                 }
             }
         })
-    }
-
-    /// Process profiles and set the profile ID in the profile context on the transaction if successful
-    #[cfg(feature = "processing")]
-    fn process_profiles(&self, state: &mut ProcessEnvelopeState) {
-        let profiling_enabled = state.project_state.has_feature(Feature::Profiling);
-        let mut found_profile_id = None;
-        state.managed_envelope.retain_items(|item| match item.ty() {
-            ItemType::Profile => {
-                if !profiling_enabled {
-                    return ItemAction::DropSilently;
-                }
-                // If we don't have an event at this stage, we need to drop the profile.
-                let Some(event) = state.event.value() else {
-                    return ItemAction::DropSilently;
-                };
-                match relay_profiling::expand_profile(&item.payload(), event) {
-                    Ok((profile_id, payload)) => {
-                        if payload.len() <= self.inner.config.max_profile_size() {
-                            found_profile_id = Some(profile_id);
-                            item.set_payload(ContentType::Json, payload);
-                            ItemAction::Keep
-                        } else {
-                            ItemAction::Drop(Outcome::Invalid(DiscardReason::Profiling(
-                                relay_profiling::discard_reason(
-                                    relay_profiling::ProfileError::ExceedSizeLimit,
-                                ),
-                            )))
-                        }
-                    }
-                    Err(err) => ItemAction::Drop(Outcome::Invalid(DiscardReason::Profiling(
-                        relay_profiling::discard_reason(err),
-                    ))),
-                }
-            }
-            _ => ItemAction::Keep,
-        });
-        if found_profile_id.is_none() {
-            // Remove profile context from event.
-            if let Some(event) = state.event.value_mut() {
-                if event.ty.value() == Some(&EventType::Transaction) {
-                    if let Some(contexts) = event.contexts.value_mut() {
-                        contexts.remove::<ProfileContext>();
-                    }
-                }
-            }
-        }
     }
 
     /// Creates and initializes the processing state.
@@ -2180,7 +2084,7 @@ impl EnvelopeProcessorService {
             self.inner.outcome_aggregator.clone(),
         );
         replay::process(state, self.inner.config.clone())?;
-        self.filter_profiles(state);
+        profile::filter(state);
 
         if state.creates_event() {
             // Some envelopes only create events in processing relays; for example, unreal events.
@@ -2192,7 +2096,7 @@ impl EnvelopeProcessorService {
             });
 
             self.extract_event(state)?;
-            self.transfer_profile_id(state);
+            profile::transfer_id(state);
 
             if_processing!({
                 self.process_unreal(state)?;
@@ -2220,7 +2124,7 @@ impl EnvelopeProcessorService {
 
         if_processing!({
             self.enforce_quotas(state)?;
-            self.process_profiles(state);
+            profile::process(state, self.inner.config.clone());
             self.process_check_ins(state);
         });
 
@@ -2731,7 +2635,6 @@ mod tests {
     use std::env;
 
     use chrono::{DateTime, TimeZone, Utc};
-    use insta::assert_debug_snapshot;
     use relay_base_schema::metrics::{DurationUnit, MetricUnit};
     use relay_common::glob2::LazyGlob;
     use relay_dynamic_config::ProjectConfig;
@@ -3038,31 +2941,6 @@ mod tests {
 
         // regression test to ensure we don't fail parsing an empty file
         result.expect("event_from_attachments");
-    }
-
-    /// Creates test processor that can be initialized in sync tests.
-    fn create_test_processor_sync(config: Config) -> EnvelopeProcessorService {
-        let inner = InnerProcessor {
-            config: Arc::new(config),
-            envelope_manager: Addr::dummy(),
-            project_cache: Addr::dummy(),
-            outcome_aggregator: Addr::dummy(),
-            upstream_relay: Addr::dummy(),
-            test_store: Addr::dummy(),
-            #[cfg(feature = "processing")]
-            rate_limiter: None,
-            #[cfg(feature = "processing")]
-            redis_pool: None,
-            geoip_lookup: None,
-            #[cfg(feature = "processing")]
-            aggregator: Addr::dummy(),
-            #[cfg(feature = "processing")]
-            metric_meta_store: None,
-        };
-
-        EnvelopeProcessorService {
-            inner: Arc::new(inner),
-        }
     }
 
     fn process_envelope_with_root_project_state(
@@ -3664,97 +3542,5 @@ mod tests {
         );
 
         assert_eq!(get_sampling_match(res).sample_rate(), 0.4);
-    }
-
-    #[test]
-    fn test_profile_id_transfered() {
-        // Setup
-        let processor = create_test_processor_sync(Default::default());
-        let event_id = EventId::new();
-        let dsn = "https://e12d836b15bb49d7bbf99e64295d995b:@sentry.io/42"
-            .parse()
-            .unwrap();
-        let request_meta = RequestMeta::new(dsn);
-        let mut envelope = Envelope::from_request(Some(event_id), request_meta);
-
-        // Add a valid transaction item.
-        envelope.add_item({
-            let mut item = Item::new(ItemType::Transaction);
-
-            item.set_payload(
-                ContentType::Json,
-                r#"
-            {
-                "type": "transaction",
-                "transaction": "/foo/",
-                "timestamp": 946684810.0,
-                "start_timestamp": 946684800.0,
-                "contexts": {
-                    "trace": {
-                    "trace_id": "4c79f60c11214eb38604f4ae0781bfb2",
-                    "span_id": "fa90fdead5f74053",
-                    "op": "http.server",
-                    "type": "trace"
-                    }
-                },
-                "transaction_info": {
-                    "source": "url"
-                }
-            }
-            "#,
-            );
-            item
-        });
-
-        // Add a profile to the same envelope.
-        envelope.add_item({
-            let mut item = Item::new(ItemType::Profile);
-            item.set_payload(
-                ContentType::Json,
-                r#"{
-                    "profile_id": "012d836b15bb49d7bbf99e64295d995b",
-                    "version": "1",
-                    "platform": "android",
-                    "os": {"name": "foo", "version": "bar"},
-                    "device": {"architecture": "zap"},
-                    "timestamp": "2023-10-10 00:00:00Z"
-                }"#,
-            );
-            item
-        });
-
-        let mut project_state = ProjectState::allowed();
-        project_state.config.features.0.insert(Feature::Profiling);
-
-        let message = ProcessEnvelope {
-            envelope: ManagedEnvelope::standalone(envelope, Addr::dummy(), Addr::dummy()),
-            project_state: Arc::new(project_state),
-            sampling_project_state: None,
-            reservoir_counters: ReservoirCounters::default(),
-            global_config: Arc::default(),
-        };
-
-        let envelope_response = processor.process(message).unwrap();
-        let ctx = envelope_response.envelope.unwrap();
-        let new_envelope = ctx.envelope();
-
-        // Get the re-serialized context.
-        let item = new_envelope
-            .get_item_by(|item| item.ty() == &ItemType::Transaction)
-            .unwrap();
-        let transaction = Annotated::<Event>::from_json_bytes(&item.payload()).unwrap();
-        let context = transaction
-            .value()
-            .unwrap()
-            .context::<ProfileContext>()
-            .unwrap();
-
-        assert_debug_snapshot!(context, @r###"
-        ProfileContext {
-            profile_id: EventId(
-                012d836b-15bb-49d7-bbf9-9e64295d995b,
-            ),
-        }
-        "###);
     }
 }
